Fix potential typo in mag-format docs
Hi there, 

Thanks so much to all the team for all the amazing work on getting OpenAlex up and running, we're really excited to start using it. 

Was reading through the docs and spotted what I think is a slight typo when discussing the differences between the new and old format.